% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_arm_BAR.R
\name{adaptiveBiasedCoin}
\alias{adaptiveBiasedCoin}
\title{Simulate a two-arm trial using  adaptive coin design}
\usage{
adaptiveBiasedCoin(potential, alpha  = 0.0, beta = 1.0, ret_prob = FALSE)
}
\arguments{

 \item{outcome}
 { A matrix with one potential outcome data for each arm. The first column is the enrolment time, the second a place holder for the arm the third and fourth potential outcome for the SOC and the experimental arm}
\item{alpha, beta}{this are design parameters}
\item{ret_prob}
{if \code{TRUE} return the probability to be assigned to the experimental arm for each of the subject}
}
\value{
The function return a matrix with the same entrance as \code{potential}. The second column includs allocation of patients to arms according to the adaptive coin design (Wei (1977, 1978))
}
\description{
This function implement the adaptive biased coin design as in Wei (1977,1978).
}

\references{
Wei LJ. (1977)
\emph{A class of designs for sequential clinical trials.}
\emph{Journal of the American Statistical Association}  72, 382--386.

Wei LJ. (1978)
\emph{The adapative biased coin design for sequential experiments.} 
\emph{Annals of Statistics} 6, 92--100.
}
