% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adj_Z.R
\name{Adj_Z}
\alias{Adj_Z}
\title{Compute an Adjusted Z statistic conditioning on the discrepancy measure T_1 -T_0}
\usage{
Adj_Z(Y_0, Y_1, T_0, T_1)
}
\arguments{
\item{Y_0}{response in the control group}

\item{Y_1}{response in the treatment group}

\item{T_0}{enrollment time (or a patient level covariate) for the control group}

\item{T_1}{enrollment time (or a patient level covariate) for the treatment group}
}
\value{
Returns a vector including the treatment effect estimate and its variance
}
\description{
Compute the adjusted statistic \eqn{\widehat \gamma_a^{\mbox{\tiny adj}}  = \widehat \gamma_a - \widehat{\mathbb E} [\widehat \gamma_a  - \gamma_a|   \Delta_a]}, where \eqn{\Delta_a = \bar{T}_a - \bar{T_0}}, is the mean difference of T between the treatment and the control group.
}
