% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_arm_BAR.R
\name{multi_arm_BAR}
\alias{multi_arm_BAR}
\title{Simulate a Trial using Bayesian response-adaptive (BAR) randomization}
\usage{
multi_arm_BAR(potential, rand, seed = NULL, omega = 0, N_min = 0,
  outcome_delay = 0)
}
\arguments{
\item{potential}{potential outcome matrix each row as
(enrollment_time,Assigned_place_holder, Y_0, ...,Y_A  )}

\item{rand}{A vector \code{rand =c(H, gamma, b)} of randomization three tuning parameter.
The first and second element correspond to the maximum
and the shap parameter of the power function \eqn{h(x) = H * (x/n)^gamma}
The last parameter controls the randomization to the control.}

\item{seed}{include a seed for the simulation.}

\item{omega}{tuning parameter for the exponential function to guarantee approximately N_min patients in each experimental arm.}

\item{N_min}{Approximate minimum for each experimental arm.}

\item{outcome_delay}{After how much time the outcome of patients \code{i} become available for the randomization procedure.}
}
\value{
The function return a matrix with the same entrance as \code{potential}. The second column includs allocation of patients to arms according to the BAR procedure.
}
\description{
Generate a multi-arm trial under BAR
}
