% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomize_BAR_add.R
\name{randomize_BAR_add}
\alias{randomize_BAR_add}
\title{Generate random numbers for a Bayesian Adaptive treatment assignment}
\usage{
randomize_BAR_add(Accrual, k.c, batch, m.k, n.k, PrPTE, rand, q.vec,
  n.a.max, Active = rep(1, length(PrPTE)), nr.a = 1)
}
\arguments{
\item{Accrual}{The number of patients randomized to each treatment arm.}

\item{k.c}{Current number of groups of added arms (initial group counts too).}

\item{batch}{Group membership for each treatment arm (for control put 1).}

\item{m.k}{Mean parameter of the Gompertz function for each treatment group,}

\item{n.k}{Extension of the overall sample size of the trial after each group of added arms.}

\item{PrPTE}{\code{PrPTE} is a vector which contains, for each experimental arm,
the posterior probability of treatment efficacy
\code{PrPTE[a]} \eqn{=p(\theta_a > \theta_0 | Data)}.
The vector can be generated with \code{\link{Posterior.with.control}}.}

\item{rand}{A vector three tuning parameter for adaptive randomization \code{rand =c(a,b,c)}.
The first and second element correspond to the maximum
and the shap parameter of the power function \eqn{h(x) = a * (x/n)^b}
The last parameter controls the randomization to the control.}

\item{q.vec}{The 3 common parameter \eqn{ (r_0, r_1,r_3) } of the Gompertz function
\eqn{q_k(x) = r_0 + r_1 \exp(-\exp( -r_2 [x-m_k] ))}.}

\item{n.a.max}{Vector of maximum sample sizes for each experimental arm.}

\item{Active}{Activity status for each experimental arm at the i-th arrival,
i.e. \code{Active[k]=1} or \code{TRUE} if  k-th arm is active.}

\item{nr.a}{How many random numbers should be generated}
}
\value{
Returns a set of \code{nr.a} treatment assignment numbers.
}
\description{
Generate a random vector of \code{nr.a} treatment assignment variables for the Bayesian adaptive design.
}
\details{
The function generates integers between 1 and A, where A is the number of arms in the trial.
Each random integer \eqn{ A_i= \{1,\ldots, A\}} denotes the assignment of patient i to
the control (\eqn{A_i=1}) or experimental arm  (\eqn{ A_i= \{2,\ldots, A\}}).
The random numbers are drawn under a Bayesian outcome-adaptive assignment model.

This function include minor changes  for the one in the AddArmToTrial; refer to \url{http://bcb.dfci.harvard.edu/∼steffen/software.html}.
}
\keyword{internal}
