% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_BAR_add_trial.R
\name{simulate_BAR_add_trial}
\alias{simulate_BAR_add_trial}
\title{Simulate a Trial using Bayesian response-adaptive (BAR) randomization}
\usage{
simulate_BAR_add_trial(Potential.Outcome, added.at = rep(1,
  ncol(Potential.Outcome)), Outcome.delay = 0, n.a.max, rand, m.k, n.k,
  q.vec, stop.par = c(f = 0, gf = 1), do.futi.stop = TRUE)
}
\arguments{
\item{Potential.Outcome}{A matrix with one potential outcome data for each arm.
The data can be generated with \code{\link{Generat.Outcome}}.}

\item{added.at}{A vector of adding times for each arm.
For all initial arms use \code{0}.}

\item{Outcome.delay}{Expected time between enrolments and avalability of the outcome.
I.e. if the endpoint is respons to treatment at 6 month, \code{Outcome.delay=6}.}

\item{n.a.max}{Vector of maximum sample sizes for each experimental arm.}

\item{rand}{A vector \code{rand =c(H, gamma, b)} of randomization three tuning parameter.
The first and second element correspond to the maximum
and the shap parameter of the power function \eqn{h(x) = H * (x/n)^gamma}
The last parameter controls the randomization to the control.}

\item{m.k}{Mean parameter of the Gompertz function for each batch}

\item{n.k}{Extension of the overall sample size for each added group}

\item{q.vec}{The 3 common parameter \eqn{ (r_0, r_1,r_3) } of the Gompertz function
\eqn{q_k(x) = r_0 + r_1 \exp(-\exp( -r_2 [x-m_k] ))}.}

\item{stop.par}{Vector of stopping parameters with default c(f=0, g=1).}

\item{do.futi.stop}{If \code{TRUE} do sequantial futility analysis.}
}
\value{
The function returns a list \code{list(TrialData = Potential.Outcome, Outcome=Outcome, Active=Active)} of three elements.

\code{TrialData} is identical to the impute matrix \code{Potential.Outcome}
except that the 2nd  column now contains treatment assignment variables, i.e.
\code{Potential.Outcome[10,2]} equals one if the 10th patients was Accrual to the control.
\code{Potential.Outcome[10,2]} equals 2 if the 10th patients was Accrual to the arm 1.

\code{Outcome} is an array of outcome data,
where \code{Outcome[,,k]} corresponds to the outcome data in the kth group of added arms.

\code{Active} corresponds to the activity status of each arm.
}
\description{
Generate a multi-arm trial under BAR whit the possibility of adding arms to the trial.
}
\details{
This function includes minor changes form the one implemented in the R package AddArmToTrial; refer to \url{http://bcb.dfci.harvard.edu/∼steffen/software.html}.
}
