% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_data.R
\name{update_data}
\alias{update_data}
\title{Update outcome data during the active accrual period and during the follow up time.}
\usage{
update_data(i, t.i, Potential.Outcome, Outcome.delay, Outcome, n.last,
  batch, N.target.0, M.k, during.accrual = TRUE)
}
\arguments{
\item{i}{current patients}

\item{t.i}{current trial time}

\item{Potential.Outcome}{matrix of potential outcome data}

\item{Outcome.delay}{Waiting time until each outcome is available}

\item{Outcome}{Array of outcome data}

\item{n.last}{First n.last patient data where already available}

\item{batch}{Vector indicating for each arm the adding-group A.k}

\item{N.target.0}{Initial target sample size for the control arm.}

\item{M.k}{Vector of unique adding times.}

\item{during.accrual}{TRUE is update during active accrual period, FALSE if update during follow up perior.}
}
\description{
Update outcome data during the active accrual period and during the follow up time.
}
\keyword{internal}
