% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Z_prop.R
\name{Z_prop}
\alias{Z_prop}
\title{Compute a standardized difference in proportion to perform Z-test}
\usage{
Z_prop(Y_0, Y_1)
}
\arguments{
\item{Y_0}{response vector for the constrol group}

\item{Y_1}{response vector for the treatment group}
}
\value{
The standardized difference in proportion
}
\description{
Compute a standardized difference in proportion to perform Z-test
}
